+++ src/ring/adapter/undertow.clj
@@ -13,28 +13,65 @@
 
 #_(set! *warn-on-reflection* true)
 
+;; TODO: cleanup
 (defn ^:no-doc undertow-handler
-  "Returns an Undertow HttpHandler implementation for the given Ring handler."
-  [handler non-blocking]
-  (reify HttpHandler
-    (handleRequest [_ exchange]
-      (when-not non-blocking
-        (.startBlocking exchange))
-      (let [request-map  (build-exchange-map exchange)
-            response-map (handler request-map)]
-        (if-let [ws-config (:undertow/websocket response-map)]
-          (->> ws-config (ws/ws-callback) (ws/ws-request exchange))
-          (set-exchange-response exchange response-map))))))
+  "Returns an function that returns Undertow HttpHandler implementation for the given Ring handler."
+  [{:keys [dispatch? websocket?]
+    :or   {dispatch?  true
+           websocket? true}}]
+  (fn [handler]
+    (reify HttpHandler
+      (handleRequest [_ exchange]
+        (when-not dispatch? (.startBlocking exchange))
+        (let [request-map  (build-exchange-map exchange)
+              response-map (handler request-map)]
+          (if websocket?
+            (if-let [ws-config (:undertow/websocket response-map)]
+              (->> ws-config (ws/ws-callback) (ws/ws-request exchange))
+              (set-exchange-response exchange response-map))
+            (set-exchange-response exchange response-map)))))))
 
-(defn ^:no-doc on-io-proxy
-  [handler]
-  (undertow-handler handler false))
+(defn ^:no-doc async-undertow-handler
+  [{:keys [websocket?]
+    :or   {websocket? true}}]
+  (fn [handler]
+    (reify HttpHandler
+      (handleRequest [_ exchange]
+        (.dispatch exchange
+                   (fn []
+                     (handler
+                       (build-exchange-map exchange)
+                       (fn [response-map]
+                         (if websocket?
+                           (if-let [ws-config (:undertow/websocket response-map)]
+                             (->> ws-config (ws/ws-callback) (ws/ws-request exchange))
+                             (set-exchange-response exchange response-map))
+                           (set-exchange-response exchange response-map)))
+                       (fn [^Throwable exception]
+                         (set-exchange-response exchange {:status 500
+                                                          :body   (.getMessage exception)})))))))))
 
-(defn ^:no-doc dispatching-proxy
-  [handler]
-  (BlockingHandler. (undertow-handler handler true)))
+(defn ^:no-doc handler!
+  [handler builder {:keys [dispatch? handler-proxy websocket? ring-async?]
+                    :or   {dispatch?   true
+                           websocket?  true
+                           ring-async? false}
+                    :as   options}]
+  (let [target-handler-proxy (cond
+                               (some? handler-proxy) handler-proxy
+                               ring-async? (async-undertow-handler options)
+                               :else (undertow-handler options))]
+    (cond->> (target-handler-proxy handler)
 
-(defn ^:no-doc tune
+             (and (nil? handler-proxy)
+                  (not ring-async?)
+                  dispatch?)
+             (BlockingHandler.)
+
+             true
+             (.setHandler builder))))
+
+(defn ^:no-doc tune!
   [^Undertow$Builder builder {:keys [io-threads worker-threads buffer-size direct-buffers?]}]
   (cond-> builder
           io-threads (.setIoThreads io-threads)
@@ -42,7 +79,7 @@
           buffer-size (.setBufferSize buffer-size)
           (not (nil? direct-buffers?)) (.setDirectBuffers direct-buffers?)))
 
-(defn ^:no-doc listen
+(defn ^:no-doc listen!
   [^Undertow$Builder builder {:keys [host port ssl-port ssl-context key-managers trust-managers]
                               :as   options
                               :or   {host "localhost" port 80 ssl-context (keystore->ssl-context options)}}]
@@ -51,7 +88,7 @@
           (and ssl-port (not ssl-context)) (.addHttpsListener ^int ssl-port ^String host ^"[Ljavax.net.ssl.KeyManager;" key-managers ^"[Ljavax.net.ssl.TrustManager;" trust-managers)
           (and port) (.addHttpListener port host)))
 
-(defn ^:no-doc client-auth [^Undertow$Builder builder {:keys [client-auth]}]
+(defn ^:no-doc client-auth! [^Undertow$Builder builder {:keys [client-auth]}]
   (when client-auth
     (case client-auth
       (:want :requested)
@@ -59,7 +96,7 @@
       (:need :required)
       (.setSocketOption builder Options/SSL_CLIENT_AUTH_MODE SslClientAuthMode/REQUIRED))))
 
-(defn ^:no-doc http2 [^Undertow$Builder builder {:keys [http2?]}]
+(defn ^:no-doc http2! [^Undertow$Builder builder {:keys [http2?]}]
   (when http2?
     (.setServerOption builder UndertowOptions/ENABLE_HTTP2 true)
     (.setServerOption builder UndertowOptions/ENABLE_SPDY true)))
@@ -84,23 +121,20 @@
   :buffer-size      - a number, defaults to 16k for modern servers
   :direct-buffers?  - boolean, defaults to true
   :dispatch?        - dispatch handlers off the I/O threads (default: true)
+  :websocket?       - built-in handler support for websocket callbacks
+  :ring-async?      - ring async support
   :handler-proxy    - an optional custom handler proxy function taking handler as single argument
 
   Returns an Undertow server instance. To stop call (.stop server)."
-  [handler {:keys [dispatch? handler-proxy]
-            :or   {dispatch? true}
-            :as   options}]
-  (let [^Undertow$Builder builder (Undertow/builder)
-        handler-proxy             (or handler-proxy
-                                      (if dispatch? dispatching-proxy on-io-proxy))]
+  [handler options]
+  (let [^Undertow$Builder builder (Undertow/builder)]
+    (handler! handler builder options)
+    (tune! builder options)
+    (http2! builder options)
+    (client-auth! builder options)
+    (listen! builder options)
 
-    (.setHandler builder (handler-proxy handler))
-    (tune builder options)
-    (http2 builder options)
-    (client-auth builder options)
-    (listen builder options)
-
-    (when-let [configurator (:configurator options)]
+    (when-some [configurator (:configurator options)]
       (configurator builder))
 
     (let [^Undertow server (.build builder)]